+++ integrations/google_genai/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [integrations/google_genai-v3.10.0] - 2026-03-10
+
+### 🚀 Features
+
+- *(google-genai)* Add timeout and max_retries to chat generator (#2875)
+- GoogleGenAIChatGenerator provides supported models list (#2930)
+- Add GoogleGenAIMultimodalDocumentEmbedder to support gemini-embedding-2 (#2944)
+
+### 🌀 Miscellaneous
+
+
 ## [integrations/google_genai-v3.8.0] - 2026-03-02
 
 ### 🚀 Features